Version 1.2
Now disposing translation units after use, greatly improves memory usage
Made `length` of `TypeInfoArray` a `int64_t` to match return type of `clang_getArraySize()`
Now using a single index for all translation units
Disabled error output for index
Now showing how many files were parsed
Switched from hardcoded size to `sizeof`
Updated docs
Fixed no namespace on template fields
Fixed `-no-empty` and `-clang` arguments not parsing correctly

+++ src/source_code.h
@@ -90,7 +90,7 @@ struct TypeInfo {
 
 struct TypeInfoArray : public TypeInfo {
     const TypeInfo* underlying;
-    int length;
+    int64_t length;
 };
 
 struct TypeInfoIndirect : public TypeInfo {
